+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 - Updated minimum optional dependency versions: astropy-healpix>=1.0.2
 
 ### Fixed
+- Fixed a bug writing UVData objects to UVFITS formats with extra phase center catalog
+entries resulted in an error on read.
+- Fixed a bug where calling `UVData._set_app_coords_helper` with extra phase center
+catalog entries resulted in an error.
 - Bug in which `correct_cable_len` defaulted to `None` instead of `True`
 for `read_mwa_corr_fits`.